refactor(bridge): extract SseEventMapper from OpenCodePluginImpl
